[SOK-16] *rename UserReducer to AuthReducer *rename methods in auth-reducer.ts -remove origin host from methods in auth-reducer.ts because base url set in backendApi
