+++ packages/client/src/store/index.ts
@@ -1,11 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit'
-import UserReducer from '@/store/reducers/user-reducer'
+import AuthReducer from '@/store/reducers/auth-reducer'
 import { useDispatch } from 'react-redux'
 
 import { combineReducers } from '@reduxjs/toolkit'
 
 const rootReducer = combineReducers({
-  UserReducer,
+  AuthReducer,
 })
 export const store = configureStore({
   reducer: rootReducer,
